qa_agent/requirements.txt
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.10
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.1
langchain-pinecone==0.0.3
pinecone-client==3.0.0
python-dotenv==1.0.0
streamlit==1.29.0
tiktoken==0.5.2
